refactor: 여러가지 리팩토링 (#170)
* headver: 31 -> 32

* refactor: HomeCollegeCard tailwind 전환

* refactor: /app/ 루트 디렉토리 간소화

* rename: 멘토 카드 관련 내용 이동
