--- src/app/home-college-cards.module.css
@@ -1,56 +1,0 @@
-/* HomeCollegeCards */
-.container {
-  overflow-x: auto;
-}
-
-.items {
-  display: flex;
-  gap: 8px;
-}
-
-/* HomeCollegeCard */
-.card {
-  position: relative;
-  width: 153px;
-  /* height: 120px; */
-}
-
-.image-wrapper {
-  position: relative;
-  width: 153px;
-}
-
-.image-wrapper::before {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  height: 120px;
-  border-radius: 8px;
-  background: linear-gradient(180deg, rgba(255, 255, 255, 0) 0%, rgba(0, 0, 0, 0.35) 75%, rgba(0, 0, 0, 0.7) 100%);
-
-  content: "";
-}
-
-.image-wrapper > img {
-  border-radius: 8px;
-  object-fit: cover;
-}
-
-.name {
-  position: absolute;
-  bottom: 9px;
-  left: 10px;
-  z-index: 10;
-
-  color: #fff;
-  font-feature-settings:
-    "liga" off,
-    "clig" off;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: 160%; /* 22.4px */
-  letter-spacing: 0.15px;
-}
